Backend/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
boto3==1.35.85
python-dotenv==1.0.1
argon2-cffi==23.1.0
PyJWT==2.9.0
requests==2.32.3
pydantic==2.10.3
